requirements.txt
pytz==2025.1
PyYAML==6.0.2